+++ source/Messaging.Tests/CimMessageAdapter/Messages/xml/BadRequestChangeOfSupplier.xml
