+++ source/Messaging.Tests/CimMessageAdapter/Messages/xml/Confirm request Change of Supplier.xml
