+++ source/Messaging.Tests/CimMessageAdapter/Messages/xml/RequestChangeOfSupplier.xml
